XRDDEV-2196 Don't duplicate service/mapper/dto
