+++ src/centralserver/admin-service/core-api/src/main/java/org/niis/xroad/cs/admin/api/service/ConfigurationSigningKeysService.java
@@ -37,6 +37,8 @@ public interface ConfigurationSigningKeysService {
 
     List<ConfigurationSigningKey> findByTokenIdentifier(String tokenIdentifier);
 
+    ConfigurationSigningKey addKey(String sourceType, String tokenId, String keyLabel);
+
     Optional<ConfigurationSigningKey> findActiveForSource(String sourceType);
 
 }
