+++ src/centralserver/admin-service/core/src/main/java/org/niis/xroad/centralserver/restapi/service/ConfigurationSigningKeysServiceImpl.java
@@ -26,26 +26,54 @@
  */
 package org.niis.xroad.centralserver.restapi.service;
 
+import ee.ria.xroad.common.identifier.ClientId;
+import ee.ria.xroad.signer.protocol.dto.KeyInfo;
+import ee.ria.xroad.signer.protocol.dto.KeyUsageInfo;
+import ee.ria.xroad.signer.protocol.dto.TokenInfo;
+
 import lombok.RequiredArgsConstructor;
+import org.niis.xroad.centralserver.restapi.service.exception.NotFoundException;
+import org.niis.xroad.centralserver.restapi.service.exception.SignerProxyException;
 import org.niis.xroad.cs.admin.api.domain.ConfigurationSigningKey;
+import org.niis.xroad.cs.admin.api.dto.KeyLabel;
+import org.niis.xroad.cs.admin.api.facade.SignerProxyFacade;
 import org.niis.xroad.cs.admin.api.service.ConfigurationSigningKeysService;
+import org.niis.xroad.cs.admin.api.service.SystemParameterService;
+import org.niis.xroad.cs.admin.api.service.TokensService;
+import org.niis.xroad.cs.admin.core.entity.ConfigurationSigningKeyEntity;
+import org.niis.xroad.cs.admin.core.entity.ConfigurationSourceEntity;
 import org.niis.xroad.cs.admin.core.entity.mapper.ConfigurationSigningKeyMapper;
 import org.niis.xroad.cs.admin.core.repository.ConfigurationSigningKeyRepository;
+import org.niis.xroad.cs.admin.core.repository.ConfigurationSourceRepository;
+import org.springframework.beans.factory.ObjectProvider;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 
+import java.time.Instant;
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.Optional;
+import java.util.UUID;
 import java.util.stream.Collectors;
+
+import static org.niis.xroad.centralserver.restapi.service.exception.ErrorMessage.CONFIGURATION_NOT_FOUND;
+import static org.niis.xroad.centralserver.restapi.service.exception.ErrorMessage.KEY_GENERATION_FAILED;
+import static org.niis.xroad.cs.admin.api.dto.PossibleAction.GENERATE_KEY;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class ConfigurationSigningKeysServiceImpl implements ConfigurationSigningKeysService {
-
+    private static final Date SIGNING_KEY_CERT_NOT_BEFORE = Date.from(Instant.EPOCH);
+    private static final Date SIGNING_KEY_CERT_NOT_AFTER = Date.from(Instant.parse("2038-01-01T00:00:00Z"));
     private final ConfigurationSigningKeyRepository configurationSigningKeyRepository;
+    private final ConfigurationSourceRepository configurationSourceRepository;
     private final ConfigurationSigningKeyMapper configurationSigningKeyMapper;
+    private final ObjectProvider<TokensService> tokensService;
+    private final SignerProxyFacade signerProxyFacade;
+    private final TokenActionsResolver tokenActionsResolver;
 
     public List<ConfigurationSigningKey> findByTokenIdentifier(String tokenIdentifier) {
         return configurationSigningKeyRepository.findByTokenIdentifier(tokenIdentifier).stream()
@@ -57,5 +85,72 @@ public class ConfigurationSigningKeysServiceImpl implements ConfigurationSigning
         // TODO pass haNodeName if HA enabled
         return configurationSigningKeyRepository.findActiveForSource(sourceType, null)
                 .map(configurationSigningKeyMapper::toTarget);
+    }
+
+    @Override
+    public ConfigurationSigningKey addKey(String sourceType, String tokenId, String keyLabel) {
+
+        var response = new ConfigurationSigningKey();
+        response.setActiveSourceSigningKey(Boolean.FALSE);
+
+        ConfigurationSourceEntity configurationSourceEntity =
+                findConfigurationSourceBySourceType(sourceType.toLowerCase());
+
+        final TokenInfo tokenInfo = tokensService.getObject().getToken(tokenId);
+        tokenActionsResolver.requireAction(GENERATE_KEY, tokenInfo);
+
+        KeyInfo keyInfo;
+        try {
+            keyInfo = signerProxyFacade.generateKey(tokenId, keyLabel);
+        } catch (Exception e) {
+            throw new SignerProxyException(KEY_GENERATION_FAILED);
+        }
+
+        final Instant generatedAt = Instant.now();
+
+        final ClientId.Conf clientId = ClientId.Conf.create(SystemParameterService.INSTANCE_IDENTIFIER,
+                "selfsigned", UUID.randomUUID().toString());
+
+        try {
+            final byte[] selfSignedCert = signerProxyFacade.generateSelfSignedCert(keyInfo.getId(), clientId,
+                    KeyUsageInfo.SIGNING,
+                    "N/A",
+                    SIGNING_KEY_CERT_NOT_BEFORE,
+                    SIGNING_KEY_CERT_NOT_AFTER);
+
+            ConfigurationSigningKeyEntity signingKey = new ConfigurationSigningKeyEntity(keyInfo.getId(),
+                    selfSignedCert, generatedAt, tokenId);
+
+            if (configurationSourceEntity.getConfigurationSigningKey() == null) {
+                configurationSourceEntity.setConfigurationSigningKey(signingKey);
+                response.setActiveSourceSigningKey(Boolean.TRUE);
+            }
+
+            configurationSourceEntity.getConfigurationSigningKeys().add(signingKey);
+
+            configurationSourceRepository.save(configurationSourceEntity);
+        } catch (Exception e) {
+            deleteKey(keyInfo.getId());
+        }
+
+        return response.setKeyIdentifier(keyInfo.getId())
+                .setLabel(new KeyLabel(keyInfo.getLabel()))
+                .setAvailable(keyInfo.isAvailable())
+                .setKeyGeneratedAt(generatedAt)
+                .setPossibleActions(new ArrayList<>(tokenActionsResolver.resolveActions(tokenInfo)))
+                .setTokenIdentifier(tokenId);
+    }
+
+    private void deleteKey(String keyId) {
+        try {
+            signerProxyFacade.deleteKey(keyId, true);
+        } catch (Exception e) {
+            throw new SignerProxyException(KEY_GENERATION_FAILED);
+        }
+    }
+
+    private ConfigurationSourceEntity findConfigurationSourceBySourceType(String sourceType) {
+        return configurationSourceRepository.findBySourceType(sourceType)
+                .orElseThrow(() -> new NotFoundException(CONFIGURATION_NOT_FOUND));
     }
 }
